creating route, and some changes in model to set a new value

+++ src/database/migrations/1614458777954-Bills.ts
@@ -22,7 +22,7 @@ export class Bills1614458777954 implements MigrationInterface {
                     },
                     {
                         name: 'value',
-                        type: 'decimal'
+                        type: 'float'
                     },
                     {
                         name: 'expiration',
